+++ src/main/java/com/dnd/runus/auth/oidc/provider/apple/dto/AppleAuthTokenResponse.java
@@ -1,16 +1,15 @@
-package com.dnd.runus.client.vo;
+package com.dnd.runus.auth.oidc.provider.apple.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public record AppleAuthTokenResponse(
-    String accessToken,
-    String expiresIn,
-    String idToken,
-    String refreshToken,
-    String tokenType,
-    String error
+        String accessToken,
+        String expiresIn,
+        String idToken,
+        String refreshToken,
+        String tokenType,
+        String error
 ) {
-
 }
