--- src/main/java/com/dnd/runus/client/config/RestClientConfig.java
@@ -1,34 +1,0 @@
-package com.dnd.runus.client.config;
-
-import com.dnd.runus.client.web.AppleAuthClientComponent;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.web.client.ClientHttpRequestFactories;
-import org.springframework.boot.web.client.ClientHttpRequestFactorySettings;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.client.ClientHttpRequestFactory;
-import org.springframework.web.client.RestClient;
-import org.springframework.web.client.support.RestClientAdapter;
-import org.springframework.web.service.invoker.HttpServiceProxyFactory;
-
-import java.time.Duration;
-
-@Configuration
-public class RestClientConfig {
-
-    @Bean
-    public AppleAuthClientComponent appleAuthClientService(@Value("${oauth.apple.public-key-url}") String url) {
-        ClientHttpRequestFactorySettings settings = ClientHttpRequestFactorySettings.DEFAULTS
-                .withReadTimeout(Duration.ofSeconds(5))
-                .withConnectTimeout(Duration.ofSeconds(10));
-        ClientHttpRequestFactory requestFactory = ClientHttpRequestFactories.get(settings);
-
-        RestClient restClient =
-                RestClient.builder().baseUrl(url).requestFactory(requestFactory).build();
-        RestClientAdapter adapter = RestClientAdapter.create(restClient);
-        HttpServiceProxyFactory factory =
-                HttpServiceProxyFactory.builderFor(adapter).build();
-
-        return factory.createClient(AppleAuthClientComponent.class);
-    }
-}
