--- src/main/java/com/dnd/runus/client/vo/OidcPublicKey.java
@@ -1,3 +1,0 @@
-package com.dnd.runus.client.vo;
-
-public record OidcPublicKey(String kty, String kid, String use, String alg, String n, String e) {}
